+++ Botyara.Core/Schedule.cs
@@ -19,21 +19,5 @@ namespace Botyara.Core
 		{
 
 		}
-
-		public static string Format(string formatWithNames, IDictionary<string, object> data)
-		{
-			var pos = 0;
-			var args = new List<object>();
-			var fmt = Regex.Replace(
-				formatWithNames, @"(?<={)[^}]+(?=})", m =>
-				{
-					var res = (pos++).ToString();
-					var tok = m.Groups[0].Value.Split(':');
-					args.Add(data[tok[0]]);
-					return tok.Length == 2 ? res + ":" + tok[1] : res;
-				}
-			);
-			return string.Format(fmt, args.ToArray());
-		}
 	}
 }
